dcp-t4(002): move factoryload and config into factory_definitions
Relocate pkg/transports/cli/factoryload and pkg/transports/cli/config to
pkg/services/factory_definitions/transports/cli/. Both render Factory
Definitions state for CLI operators and import only that service among product
services, so they belong beside cobracompletion in the owning service's
transports subtree.

Pure relocation: file contents are unchanged apart from import paths in the
external importers and the self-import in factoryload's external test. Import
aliases already described the package rather than its location, so they are
unchanged.

cmd/pkgboundarycheck/transport_behavior_test.go built its named-factory
path-policy fixture at the literal path pkg/transports/cli/factoryload, which
no longer exists. The fixture is synthetic and only needs a package under the
composition-root CLI tree, so it now uses pkg/transports/cli/factory, a
package this lane leaves in place. The rule under test is unchanged: the real
factoryload never called factorydefinitions.MapDir, so nothing escapes the
named-factory path policy by moving.
