+++ pkg/services/factory_definitions/transports/cli/factoryload/blocking_factory_load_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	"github.com/portpowered/infinite-you/pkg/transports/cli/factoryload"
+	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/transports/cli/factoryload"
 	configdiagnostics "github.com/portpowered/infinite-you/pkg/transports/mapping/factoryconfig/diagnostics"
 )
