+++ pkg/services/factory_definitions/transports/cli/factoryload/operator_error.go
