+++ pkg/transports/cli/root_run_selection_test.go
@@ -7,7 +7,7 @@ import (
 
 	startupcli "github.com/portpowered/infinite-you/pkg/initializer/process"
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	"github.com/portpowered/infinite-you/pkg/transports/cli/factoryload"
+	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/transports/cli/factoryload"
 	runcli "github.com/portpowered/infinite-you/pkg/transports/cli/run"
 )
